fixed no_fail

+++ bbconf/config_parser/bedbaseconfig.py
@@ -240,6 +240,8 @@ class BedBaseConfig:
             raise BedbaseS3ConnectionError(
                 "Could not upload file to s3. Connection error."
             )
+        if not os.path.exists(file_path):
+            raise BedBaseConfError(f"File {file_path} does not exist.")
         return self._boto3_client.upload_file(file_path, self.config.s3.bucket, s3_path)
 
     def upload_files_s3(
